demo_light/requirements.txt
# Agentic Reviewer Demo - Minimal Dependencies
pandas>=2.0.0
pyyaml>=6.0
aiohttp>=3.8.0

